barbershop/requirements.txt
bcrypt==4.1.3
blinker==1.8.2
click==8.1.7
DateTime==5.5
Flask==3.0.3
flask-sqlalchemy>=3.0.0
psycopg2>=2.9.4
gunicorn>=20.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pytz==2024.1
setuptools==70.0.0
Werkzeug==3.0.3
zope.interface==6.4.post2
